Move small test back to printers_test.go
